Page_1/requirements.txt
pdfplumber
pandas
reportlab
openai>=1.0.0
python-dotenv
